правки по ревью

+++ ewm-service/src/main/java/ru/practicum/service/user/UserServiceImpl.java
@@ -5,13 +5,13 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import ru.practicum.exceptions.ModelNotFoundException;
 import ru.practicum.mapper.MapperUtil;
 import ru.practicum.model.user.User;
 import ru.practicum.model.user.dto.UserCreationDto;
 import ru.practicum.model.user.dto.UserDto;
 import ru.practicum.repository.UserRepository;
 
-import java.util.Arrays;
 import java.util.List;
 
 @Service
@@ -28,15 +28,19 @@ public class UserServiceImpl implements UserService {
 
     @Override
     @Transactional(readOnly = true)
-    public List<UserDto> getAllUsers(Long[] ids, int from, int size) {
-        List<Long> usersId = null;
-        if (ids != null && ids.length != 0) usersId = Arrays.asList(ids);
+    public List<UserDto> getAllUsers(List<Long> ids, int from, int size) {
         Pageable pageable = PageRequest.of(from / size, size);
-        return userRepository.findAllById(usersId, pageable);
+        List<User> users;
+        if (ids != null) users = userRepository.findByIdIn(ids, pageable);
+        else users = userRepository.findAll(pageable).getContent();
+        return MapperUtil.convertList(users, MapperUtil::convertToUserDto);
     }
 
     @Override
     public void deleteUserById(long userId) {
+        if (!userRepository.existsById(userId)) {
+            throw new ModelNotFoundException("User with id=" + userId + " was not found");
+        }
         userRepository.deleteById(userId);
     }
 }
